requirements.txt
PyMuPDF==1.23.8
langchain==0.1.9
sentence-transformers==2.5.1
chromadb==0.4.22
openai==0.28.0
numpy==1.26.4
streamlit==1.31.1
pdfplumber==0.10.3
pdfminer.six==20221105
nltk==3.8.1
python-dotenv==1.0.1
faiss-cpu>=1.7.4
httpx>=0.26.0
